[200~Various Bug Fixes.
Added new ActionTask, the ActionTask_Sequence. This is similar to ActionTask_Composite with the exception that it will execute the tasks one after the other, in order. ActionTask_Composite was also updated to support some control flow and allowing inifinite running actions tasks (such as a looping animation) until all non-ignored completion signals are received. Updated the Environment Query to use the appriorate UObject... but why it takes such a base class instead of the one it actually wants is a mystery to me. There must have been a change in 5.0 where the handles are auto-disposing once going out of scope and no longer cached by the BrainComponent, so had to add some management there. May need to consider relying on that wierd BT_Task memory stuff so not every 'stateful' action task needs a map or some other kind of lookup. Perhaps a manager subsystem would do well here?

Fixed various deprecation warnings.

Added GetDebugInfoString for the GameplayDubugger tools overlay. Of course it displays under BehaviorTree... when BrainComponent seemed too generic to be tied to that. But it works and makes plans less of a mystery when they are scheduled.

Memory enhancements for Planner_AStar that reuses a single state for future prediction and only creates a copy when it ends up in the open list. Also, all of the objects created are parented to the Start.State and that gets destroyed after planning is complete.

Added static assertions to state template functions to avoid runtime errors of the default false.~